+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/ImageListNodeOperateHandlerImpl.java
@@ -365,7 +365,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。
@@ -481,7 +482,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。
@@ -599,7 +601,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。
@@ -710,7 +713,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。
@@ -824,7 +828,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。
@@ -952,7 +957,8 @@ public class ImageListNodeOperateHandlerImpl implements ImageListNodeOperateHand
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片列表节点";
                 settingNode = new SettingNode(
-                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE_LIST, new Date(), remark,
+                        true, category, args
                 );
             }
             // 否则更新属性。