+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/ImageNodeOperateHandlerImpl.java
@@ -257,7 +257,10 @@ public class ImageNodeOperateHandlerImpl implements ImageNodeOperateHandler {
             // 如果节点设置不存在，则初始化节点设置实体。
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片节点";
-                settingNode = new SettingNode(settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE, new Date(), remark);
+                settingNode = new SettingNode(
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE, new Date(), remark,
+                        true, category, args
+                );
             }
             // 否则更新属性。
             else {
@@ -315,7 +318,10 @@ public class ImageNodeOperateHandlerImpl implements ImageNodeOperateHandler {
             // 如果节点设置不存在，则初始化节点设置实体。
             if (Objects.isNull(settingNode)) {
                 String remark = "由 settingrepo 自动生成的图片节点";
-                settingNode = new SettingNode(settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE, new Date(), remark);
+                settingNode = new SettingNode(
+                        settingNodeKey, Constants.SETTING_NODE_TYPE_IMAGE, new Date(), remark,
+                        true, category, args
+                );
             }
             // 否则更新属性。
             else {