+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/SettingNodeMaintainServiceImplTest.java
@@ -1,6 +1,8 @@
 package com.dwarfeng.settingrepo.impl.service;
 
+import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory;
 import com.dwarfeng.settingrepo.stack.bean.entity.SettingNode;
+import com.dwarfeng.settingrepo.stack.service.SettingCategoryMaintainService;
 import com.dwarfeng.settingrepo.stack.service.SettingNodeMaintainService;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import org.apache.commons.beanutils.BeanUtils;
@@ -15,24 +17,34 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = "classpath:spring/application-context*.xml")
 public class SettingNodeMaintainServiceImplTest {
 
+    private static final String SETTING_CATEGORY_ID = "test.setting_category";
+
+    @Autowired
+    private SettingCategoryMaintainService settingCategoryMaintainService;
     @Autowired
     private SettingNodeMaintainService settingNodeMaintainService;
 
+    private SettingCategory settingCategory;
     private List<SettingNode> settingNodes;
 
     @Before
     public void setUp() {
+        settingCategory = new SettingCategory(
+                new StringIdKey(SETTING_CATEGORY_ID), "formatterType", "formatterParam", "remark"
+        );
         settingNodes = new ArrayList<>();
         for (int i = 0; i < 5; i++) {
             SettingNode settingNode = new SettingNode(
-                    new StringIdKey("test.setting_node." + i), 12450, new Date(), "remark"
+                    new StringIdKey("test.setting_node." + i), 12450, new Date(), "remark",
+                    true, "test.setting_category", new String[]{"foo", "bar"}
             );
             settingNodes.add(settingNode);
         }
@@ -46,6 +58,7 @@ public class SettingNodeMaintainServiceImplTest {
     @Test
     public void testForCrud() throws Exception {
         try {
+            settingCategoryMaintainService.insertOrUpdate(settingCategory);
             for (SettingNode settingNode : settingNodes) {
                 settingNodeMaintainService.insertOrUpdate(settingNode);
                 SettingNode testSettingNode = settingNodeMaintainService.get(settingNode.getKey());
@@ -58,6 +71,36 @@ public class SettingNodeMaintainServiceImplTest {
             for (SettingNode settingNode : settingNodes) {
                 settingNodeMaintainService.deleteIfExists(settingNode.getKey());
             }
+            settingCategoryMaintainService.deleteIfExists(settingCategory.getKey());
+        }
+    }
+
+    public void testForSettingCategoryCascade() throws Exception {
+        try {
+            settingCategoryMaintainService.insertOrUpdate(settingCategory);
+            for (SettingNode settingNode : settingNodes) {
+                settingNodeMaintainService.insertOrUpdate(settingNode);
+            }
+
+            assertTrue(settingNodeMaintainService.allExists(
+                    settingNodes.stream().map(SettingNode::getKey).collect(Collectors.toList()))
+            );
+
+            settingCategoryMaintainService.deleteIfExists(settingCategory.getKey());
+
+            assertTrue(settingNodeMaintainService.allExists(
+                    settingNodes.stream().map(SettingNode::getKey).collect(Collectors.toList()))
+            );
+
+            for (SettingNode settingNode : settingNodes) {
+                SettingNode testSettingNode = settingNodeMaintainService.get(settingNode.getKey());
+                assertFalse(testSettingNode.isReachable());
+            }
+        } finally {
+            for (SettingNode settingNode : settingNodes) {
+                settingNodeMaintainService.deleteIfExists(settingNode.getKey());
+            }
+            settingCategoryMaintainService.deleteIfExists(settingCategory.getKey());
         }
     }
 }